fix: download ics + move agenda into own component
